move fusion strategies in their own file
